added navigation functionality, layering code, mock data
