+++ lib/data/repositories/image_repository.dart
@@ -1,0 +1,5 @@
+import 'package:surf_flutter_summer_school_24/domain/models/advanced_image.dart';
+
+abstract interface class ImageRepository {
+  Future<List<AdvancedImage>> getImages();
+}
