+++ lib/data/repositories/mock_image_repository.dart
@@ -1,11 +1,11 @@
-import 'package:surf_flutter_summer_school_24/domain/model/advanced_image.dart';
-import 'package:surf_flutter_summer_school_24/data/image_repository.dart';
+import 'package:surf_flutter_summer_school_24/domain/models/advanced_image.dart';
+import 'package:surf_flutter_summer_school_24/data/repositories/image_repository.dart';
 
 class MockImageRepository implements ImageRepository {
   int _idCounter = 1;
 
   @override
-  Future<List<AdvancedImage>> getImage() async {
+  Future<List<AdvancedImage>> getImages() async {
     return [
       AdvancedImage(
           url: "https://basik.ru/images/nature_landscape_9/30_nature.jpg",
