+++ lib/di/theme_inherited.dart
@@ -1,4 +1,4 @@
-import 'package:surf_flutter_summer_school_24/domain/model/theme_controller.dart';
+import 'package:surf_flutter_summer_school_24/domain/models/theme_controller.dart';
 import 'package:flutter/material.dart';
 
 class ThemeInherited extends InheritedWidget {
