+++ lib/domain/models/advanced_image.dart
