+++ lib/main.dart
@@ -1,14 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:surf_flutter_summer_school_24/data/mock_image_repository.dart';
-import 'package:surf_flutter_summer_school_24/data/theme_repository.dart';
-import 'package:surf_flutter_summer_school_24/domain/interactors/advanced_image_interactor.dart';
-import 'package:surf_flutter_summer_school_24/domain/model/theme_controller.dart';
+import 'package:surf_flutter_summer_school_24/data/repositories/theme_repository.dart';
+import 'package:surf_flutter_summer_school_24/domain/models/theme_controller.dart';
 import 'package:surf_flutter_summer_school_24/di/theme_inherited.dart';
-import 'package:surf_flutter_summer_school_24/ui/image_carousel_screen.dart';
-import 'package:surf_flutter_summer_school_24/ui/main_screen.dart';
-import 'package:surf_flutter_summer_school_24/storage/theme/theme_storage.dart';
-import 'package:surf_flutter_summer_school_24/ui/theme_builder.dart';
+import 'package:surf_flutter_summer_school_24/ui/screens/main_screen.dart';
+import 'package:surf_flutter_summer_school_24/ui/widgets/theme_builder.dart';
+import 'package:surf_flutter_summer_school_24/data/storage/theme_storage.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -24,7 +21,6 @@ void main() async {
       child: const MyApp(),
     ),
   );
-
 }
 
 class MyApp extends StatelessWidget {
@@ -39,7 +35,7 @@ class MyApp extends StatelessWidget {
           themeMode: themeMode,
           theme: ThemeData.light(),
           darkTheme: ThemeData.dark(),
-          home: ImageCarouselScreen(),
+          home: const MainScreen(),
         );
       },
     );
